+++ src/main/java/com/refinedmods/refinedstorage/mekanism/grid/ChemicalGridExtractionStrategy.java
@@ -1,7 +1,7 @@
 package com.refinedmods.refinedstorage.mekanism.grid;
 
-import com.refinedmods.refinedstorage.api.grid.operations.GridExtractMode;
-import com.refinedmods.refinedstorage.api.grid.operations.GridOperations;
+import com.refinedmods.refinedstorage.api.network.node.grid.GridExtractMode;
+import com.refinedmods.refinedstorage.api.network.node.grid.GridOperations;
 import com.refinedmods.refinedstorage.common.api.grid.Grid;
 import com.refinedmods.refinedstorage.common.api.grid.strategy.GridExtractionStrategy;
 import com.refinedmods.refinedstorage.common.api.support.resource.PlatformResourceKey;