+++ CHANGELOG.md
@@ -7,6 +7,12 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [1.0.0] - 2025-03-28
+
+### Fixed
+
+-   Support for Refined Storage v2.0.0-beta.1.
+
 ## [0.3.0] - 2025-02-23
 
 ### Changed
@@ -43,7 +49,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 -   Chemical support in the External Storage.
 -   Ghost dragging support for chemicals in JEI and EMI.
 
-[Unreleased]: https://github.com/refinedmods/refinedstorage-mekanism-integration/compare/v0.3.0...HEAD
+[Unreleased]: https://github.com/refinedmods/refinedstorage-mekanism-integration/compare/v1.0.0...HEAD
+
+[1.0.0]: https://github.com/refinedmods/refinedstorage-mekanism-integration/compare/v0.3.0...v1.0.0
 
 [0.3.0]: https://github.com/refinedmods/refinedstorage-mekanism-integration/compare/v0.2.0...v0.3.0